refactor: Message 정적 클래스 분리
