+++ src/main/java/com/devcourse/springbootbasic/application/dto/VoucherDto.java
@@ -1,0 +1,7 @@
+package com.devcourse.springbootbasic.application.dto;
+
+public record VoucherDto(
+        VoucherType voucherType,
+        double discountValue
+) {
+}
