+++ src/main/java/com/devcourse/springbootbasic/application/repository/VoucherRepository.java
@@ -1,12 +1,12 @@
-package com.devcourse.springbootbasic.engine.voucher.repository;
+package com.devcourse.springbootbasic.application.repository;
 
-import com.devcourse.springbootbasic.engine.voucher.domain.Voucher;
+import com.devcourse.springbootbasic.application.domain.Voucher;
 
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
-// standard interface
+
 public interface VoucherRepository {
     Optional<Voucher> insert(Voucher voucher);
