--- src/main/java/com/devcourse/springbootbasic/engine/voucher/domain/VoucherDto.java
@@ -1,9 +1,0 @@
-package com.devcourse.springbootbasic.engine.voucher.domain;
-
-import com.devcourse.springbootbasic.engine.model.VoucherType;
-
-public record VoucherDto(
-        VoucherType voucherType,
-        double discountValue
-) {
-}
