--- src/main/java/com/devcourse/springbootbasic/engine/voucher/factory/VoucherFactory.java
@@ -1,7 +1,0 @@
-package com.devcourse.springbootbasic.engine.voucher.factory;
-
-import com.devcourse.springbootbasic.engine.voucher.domain.Voucher;
-
-public interface VoucherFactory {
-    Voucher create(double voucherDiscount);
-}
