+++ src/main/resources/voucher_record.txt
@@ -5,3 +5,8 @@ PERCENT_DISCOUNT -> id:1c560c17-bc6a-4cf3-a007-d71b5e63d2f7, discountPercent:5.0
 FIXED_AMOUNT -> id:d8c52b16-0cca-40d3-a8f1-423e8321f521, discountAmount:10.0
 PERCENT_DISCOUNT -> id:c77b60d6-8517-409a-9997-eeb7d933f260, discountPercent:4.0
 FIXED_AMOUNT -> id:e8a33fcb-a886-49ea-abd7-c5cc0a5d99e4, discountAmount:1.0
+FIXED_AMOUNT -> id:1ba6ba4c-25d0-4eb7-af68-d2d48996fc7a, discountFixed Amount Discount Voucher:20.0
+PERCENT_DISCOUNT(id: d0790198-3e84-47e8-82ca-8eb4c5b624cd, Percent Discount Voucher: 12)
+FIXED_AMOUNT(id: 8d3cc8e9-2c67-4cb7-9460-7d6a8f68377b, Fixed Amount Discount Voucher: 2)
+FIXED_AMOUNT(id: dea29db4-3222-4c89-8054-f1ffec3bd87b, Fixed Amount Discount Voucher: 10)
+PERCENT_DISCOUNT(id: 162fd096-50d8-4f8e-bd80-9081646e7960, Percent Discount Voucher: 32)
